Create Main - CM and GUI
This is main function which consists of the Committee Machine - Combines the Machine Learning techniques to give the final prediction - and the GUI of the simulated city. GUI is done using graphics.py package. Also consists of an all pair shortest path code to find the minimum path based on the fastest route.